refactor(core): Windows 依存を TBird.Windows へ分離、IMessageService に Warn 追加
- TBird.Core (netstandard2.0) から System.Management 依存を撤去
- Win32Methods / Win32Messages / Win32ShutdownFlags を新規 TBird.Windows
  (net8.0-windows) に移管、Win32ShowWindowStates は参照ゼロのため削除
- TBird.Core 側 FileUtil / DirectoryUtil は GetShortPathName を内部で
  直接 DllImport 宣言（属性は kernel32.dll のみ、属性差分なし）
- TBird.Wpf に TBird.Windows ProjectReference を追加、WpfDialog.cs に
  using TBird.Windows; を追加
- IMessageService / MessageService に Warn メソッドを追加（Info 直後）
- MessageType enum に Warn 値を追加（Info の直後に挿入、既存序数を保持）
- ConsoleMessageService に virtual Warn 実装を追加
- ConsoleMessageService.GetString を protected virtual に格上げ、内部の
  AppendLogfile 副作用を Error / Exception メソッド直下に移動（純粋関数化）
- TBird.Library.sln に TBird.Windows を追加（Libraries フォルダ配下）

TBird.Core の TFM は netstandard2.0 のまま据え置き。WPF 系 app-* ブランチ
の開発者環境に MAUI workload 等の追加要件は発生しない。
