+++ TBird.Core/TBird.Core.csproj
@@ -8,7 +8,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="System.Management" Version="10.0.1" />
   </ItemGroup>
 
 </Project>
