--- TBird.Core/_ROOT/Win32Messages.cs
@@ -1,1005 +1,0 @@
-﻿namespace TBird.Core
-{
-	public static class Win32Messages
-	{
-		public const int WM_NULL = 0x0000;
-		public const int WM_CREATE = 0x0001;
-		public const int WM_DESTROY = 0x0002;
-		public const int WM_MOVE = 0x0003;
-		public const int WM_SIZE = 0x0005;
-		public const int WM_ACTIVATE = 0x0006;
-		public const int WM_SETFOCUS = 0x0007;
-		public const int WM_KILLFOCUS = 0x0008;
-		public const int WM_ENABLE = 0x000a;
-		public const int WM_SETREDRAW = 0x000b;
-		public const int WM_SETTEXT = 0x000c;
-		public const int WM_GETTEXT = 0x000d;
-		public const int WM_GETTEXTLENGTH = 0x000e;
-		public const int WM_PAINT = 0x000f;
-		public const int WM_CLOSE = 0x0010;
-		public const int WM_QUERYENDSESSION = 0x0011;
-		public const int WM_QUIT = 0x0012;
-		public const int WM_QUERYOPEN = 0x0013;
-		public const int WM_ERASEBKGND = 0x0014;
-		public const int WM_SYSCOLORCHANGE = 0x0015;
-		public const int WM_ENDSESSION = 0x0016;
-		public const int WM_SHOWWINDOW = 0x0018;
-		public const int WM_CTLCOLOR = 0x0019;
-		public const int WM_WININICHANGE = 0x001a;
-		public const int WM_DEVMODECHANGE = 0x001b;
-		public const int WM_ACTIVATEAPP = 0x001c;
-		public const int WM_FONTCHANGE = 0x001d;
-		public const int WM_TIMECHANGE = 0x001e;
-		public const int WM_CANCELMODE = 0x001f;
-		public const int WM_SETCURSOR = 0x0020;
-		public const int WM_MOUSEACTIVATE = 0x0021;
-		public const int WM_CHILDACTIVATE = 0x0022;
-		public const int WM_QUEUESYNC = 0x0023;
-		public const int WM_GETMINMAXINFO = 0x0024;
-		public const int WM_PAINTICON = 0x0026;
-		public const int WM_ICONERASEBKGND = 0x0027;
-		public const int WM_NEXTDLGCTL = 0x0028;
-		public const int WM_SPOOLERSTATUS = 0x002a;
-		public const int WM_DRAWITEM = 0x002b;
-		public const int WM_MEASUREITEM = 0x002c;
-		public const int WM_DELETEITEM = 0x002d;
-		public const int WM_VKEYTOITEM = 0x002e;
-		public const int WM_CHARTOITEM = 0x002f;
-		public const int WM_SETFONT = 0x0030;
-		public const int WM_GETFONT = 0x0031;
-		public const int WM_SETHOTKEY = 0x0032;
-		public const int WM_GETHOTKEY = 0x0033;
-		public const int WM_QUERYDRAGICON = 0x0037;
-		public const int WM_COMPAREITEM = 0x0039;
-		public const int WM_GETOBJECT = 0x003d;
-		public const int WM_COMPACTING = 0x0041;
-		public const int WM_COMMNOTIFY = 0x0044;
-		public const int WM_WINDOWPOSCHANGING = 0x0046;
-		public const int WM_WINDOWPOSCHANGED = 0x0047;
-		public const int WM_POWER = 0x0048;
-		public const int WM_COPYGLOBALDATA = 0x0049;
-		public const int WM_COPYDATA = 0x004a;
-		public const int WM_CANCELJOURNAL = 0x004b;
-		public const int WM_NOTIFY = 0x004e;
-		public const int WM_INPUTLANGCHANGEREQUEST = 0x0050;
-		public const int WM_INPUTLANGCHANGE = 0x0051;
-		public const int WM_TCARD = 0x0052;
-		public const int WM_HELP = 0x0053;
-		public const int WM_USERCHANGED = 0x0054;
-		public const int WM_NOTIFYFORMAT = 0x0055;
-		public const int WM_CONTEXTMENU = 0x007b;
-		public const int WM_STYLECHANGING = 0x007c;
-		public const int WM_STYLECHANGED = 0x007d;
-		public const int WM_DISPLAYCHANGE = 0x007e;
-		public const int WM_GETICON = 0x007f;
-		public const int WM_SETICON = 0x0080;
-		public const int WM_NCCREATE = 0x0081;
-		public const int WM_NCDESTROY = 0x0082;
-		public const int WM_NCCALCSIZE = 0x0083;
-		public const int WM_NCHITTEST = 0x0084;
-		public const int WM_NCPAINT = 0x0085;
-		public const int WM_NCACTIVATE = 0x0086;
-		public const int WM_GETDLGCODE = 0x0087;
-		public const int WM_SYNCPAINT = 0x0088;
-		public const int WM_NCMOUSEMOVE = 0x00a0;
-		public const int WM_NCLBUTTONDOWN = 0x00a1;
-		public const int WM_NCLBUTTONUP = 0x00a2;
-		public const int WM_NCLBUTTONDBLCLK = 0x00a3;
-		public const int WM_NCRBUTTONDOWN = 0x00a4;
-		public const int WM_NCRBUTTONUP = 0x00a5;
-		public const int WM_NCRBUTTONDBLCLK = 0x00a6;
-		public const int WM_NCMBUTTONDOWN = 0x00a7;
-		public const int WM_NCMBUTTONUP = 0x00a8;
-		public const int WM_NCMBUTTONDBLCLK = 0x00a9;
-		public const int WM_NCXBUTTONDOWN = 0x00ab;
-		public const int WM_NCXBUTTONUP = 0x00ac;
-		public const int WM_NCXBUTTONDBLCLK = 0x00ad;
-		public const int WM_INPUT = 0x00ff;
-		public const int WM_KEYDOWN = 0x0100;
-		public const int WM_KEYFIRST = 0x0100;
-		public const int WM_KEYUP = 0x0101;
-		public const int WM_CHAR = 0x0102;
-		public const int WM_DEADCHAR = 0x0103;
-		public const int WM_SYSKEYDOWN = 0x0104;
-		public const int WM_SYSKEYUP = 0x0105;
-		public const int WM_SYSCHAR = 0x0106;
-		public const int WM_SYSDEADCHAR = 0x0107;
-		public const int WM_UNICHAR = 0x0109;
-		public const int WM_KEYLAST = 0x0109;
-		public const int WM_WNT_CONVERTREQUESTEX = 0x0109;
-		public const int WM_CONVERTREQUEST = 0x010a;
-		public const int WM_CONVERTRESULT = 0x010b;
-		public const int WM_INTERIM = 0x010c;
-		public const int WM_IME_STARTCOMPOSITION = 0x010d;
-		public const int WM_IME_ENDCOMPOSITION = 0x010e;
-		public const int WM_IME_COMPOSITION = 0x010f;
-		public const int WM_IME_KEYLAST = 0x010f;
-		public const int WM_INITDIALOG = 0x0110;
-		public const int WM_COMMAND = 0x0111;
-		public const int WM_SYSCOMMAND = 0x0112;
-		public const int WM_TIMER = 0x0113;
-		public const int WM_HSCROLL = 0x0114;
-		public const int WM_VSCROLL = 0x0115;
-		public const int WM_INITMENU = 0x0116;
-		public const int WM_INITMENUPOPUP = 0x0117;
-		public const int WM_SYSTIMER = 0x0118;
-		public const int WM_MENUSELECT = 0x011f;
-		public const int WM_MENUCHAR = 0x0120;
-		public const int WM_ENTERIDLE = 0x0121;
-		public const int WM_MENURBUTTONUP = 0x0122;
-		public const int WM_MENUDRAG = 0x0123;
-		public const int WM_MENUGETOBJECT = 0x0124;
-		public const int WM_UNINITMENUPOPUP = 0x0125;
-		public const int WM_MENUCOMMAND = 0x0126;
-		public const int WM_CHANGEUISTATE = 0x0127;
-		public const int WM_UPDATEUISTATE = 0x0128;
-		public const int WM_QUERYUISTATE = 0x0129;
-		public const int WM_CTLCOLORMSGBOX = 0x0132;
-		public const int WM_CTLCOLOREDIT = 0x0133;
-		public const int WM_CTLCOLORLISTBOX = 0x0134;
-		public const int WM_CTLCOLORBTN = 0x0135;
-		public const int WM_CTLCOLORDLG = 0x0136;
-		public const int WM_CTLCOLORSCROLLBAR = 0x0137;
-		public const int WM_CTLCOLORSTATIC = 0x0138;
-		public const int WM_MOUSEFIRST = 0x0200;
-		public const int WM_MOUSEMOVE = 0x0200;
-		public const int WM_LBUTTONDOWN = 0x0201;
-		public const int WM_LBUTTONUP = 0x0202;
-		public const int WM_LBUTTONDBLCLK = 0x0203;
-		public const int WM_RBUTTONDOWN = 0x0204;
-		public const int WM_RBUTTONUP = 0x0205;
-		public const int WM_RBUTTONDBLCLK = 0x0206;
-		public const int WM_MBUTTONDOWN = 0x0207;
-		public const int WM_MBUTTONUP = 0x0208;
-		public const int WM_MBUTTONDBLCLK = 0x0209;
-		public const int WM_MOUSELAST = 0x0209;
-		public const int WM_MOUSEWHEEL = 0x020a;
-		public const int WM_XBUTTONDOWN = 0x020b;
-		public const int WM_XBUTTONUP = 0x020c;
-		public const int WM_XBUTTONDBLCLK = 0x020d;
-		public const int WM_MOUSEHWHEEL = 0x020e;
-		public const int WM_PARENTNOTIFY = 0x0210;
-		public const int WM_ENTERMENULOOP = 0x0211;
-		public const int WM_EXITMENULOOP = 0x0212;
-		public const int WM_NEXTMENU = 0x0213;
-		public const int WM_SIZING = 0x0214;
-		public const int WM_CAPTURECHANGED = 0x0215;
-		public const int WM_MOVING = 0x0216;
-		public const int WM_POWERBROADCAST = 0x0218;
-		public const int WM_DEVICECHANGE = 0x0219;
-		public const int WM_MDICREATE = 0x0220;
-		public const int WM_MDIDESTROY = 0x0221;
-		public const int WM_MDIACTIVATE = 0x0222;
-		public const int WM_MDIRESTORE = 0x0223;
-		public const int WM_MDINEXT = 0x0224;
-		public const int WM_MDIMAXIMIZE = 0x0225;
-		public const int WM_MDITILE = 0x0226;
-		public const int WM_MDICASCADE = 0x0227;
-		public const int WM_MDIICONARRANGE = 0x0228;
-		public const int WM_MDIGETACTIVE = 0x0229;
-		public const int WM_MDISETMENU = 0x0230;
-		public const int WM_ENTERSIZEMOVE = 0x0231;
-		public const int WM_EXITSIZEMOVE = 0x0232;
-		public const int WM_DROPFILES = 0x0233;
-		public const int WM_MDIREFRESHMENU = 0x0234;
-		public const int WM_IME_REPORT = 0x0280;
-		public const int WM_IME_SETCONTEXT = 0x0281;
-		public const int WM_IME_NOTIFY = 0x0282;
-		public const int WM_IME_CONTROL = 0x0283;
-		public const int WM_IME_COMPOSITIONFULL = 0x0284;
-		public const int WM_IME_SELECT = 0x0285;
-		public const int WM_IME_CHAR = 0x0286;
-		public const int WM_IME_REQUEST = 0x0288;
-		public const int WM_IME_KEYDOWN = 0x0290;
-		public const int WM_IMEKEYDOWN = 0x0290;
-		public const int WM_IME_KEYUP = 0x0291;
-		public const int WM_IMEKEYUP = 0x0291;
-		public const int WM_NCMOUSEHOVER = 0x02a0;
-		public const int WM_MOUSEHOVER = 0x02a1;
-		public const int WM_NCMOUSELEAVE = 0x02a2;
-		public const int WM_MOUSELEAVE = 0x02a3;
-		public const int WM_CUT = 0x0300;
-		public const int WM_COPY = 0x0301;
-		public const int WM_PASTE = 0x0302;
-		public const int WM_CLEAR = 0x0303;
-		public const int WM_UNDO = 0x0304;
-		public const int WM_RENDERFORMAT = 0x0305;
-		public const int WM_RENDERALLFORMATS = 0x0306;
-		public const int WM_DESTROYCLIPBOARD = 0x0307;
-		public const int WM_DRAWCLIPBOARD = 0x0308;
-		public const int WM_PAINTCLIPBOARD = 0x0309;
-		public const int WM_VSCROLLCLIPBOARD = 0x030a;
-		public const int WM_SIZECLIPBOARD = 0x030b;
-		public const int WM_ASKCBFORMATNAME = 0x030c;
-		public const int WM_CHANGECBCHAIN = 0x030d;
-		public const int WM_HSCROLLCLIPBOARD = 0x030e;
-		public const int WM_QUERYNEWPALETTE = 0x030f;
-		public const int WM_PALETTEISCHANGING = 0x0310;
-		public const int WM_PALETTECHANGED = 0x0311;
-		public const int WM_HOTKEY = 0x0312;
-		public const int WM_PRINT = 0x0317;
-		public const int WM_PRINTCLIENT = 0x0318;
-		public const int WM_APPCOMMAND = 0x0319;
-		public const int WM_HANDHELDFIRST = 0x0358;
-		public const int WM_HANDHELDLAST = 0x035f;
-		public const int WM_AFXFIRST = 0x0360;
-		public const int WM_AFXLAST = 0x037f;
-		public const int WM_PENWINFIRST = 0x0380;
-		public const int WM_RCRESULT = 0x0381;
-		public const int WM_HOOKRCRESULT = 0x0382;
-		public const int WM_GLOBALRCCHANGE = 0x0383;
-		public const int WM_PENMISCINFO = 0x0383;
-		public const int WM_SKB = 0x0384;
-		public const int WM_HEDITCTL = 0x0385;
-		public const int WM_PENCTL = 0x0385;
-		public const int WM_PENMISC = 0x0386;
-		public const int WM_CTLINIT = 0x0387;
-		public const int WM_PENEVENT = 0x0388;
-		public const int WM_PENWINLAST = 0x038f;
-		public const int WM_PSD_PAGESETUPDLG = 0x0400;
-		public const int WM_USER = 0x0400;
-		public const int WM_CHOOSEFONT_GETLOGFONT = 0x0401;
-		public const int WM_PSD_FULLPAGERECT = 0x0401;
-		public const int WM_PSD_MINMARGINRECT = 0x0402;
-		public const int WM_PSD_MARGINRECT = 0x0403;
-		public const int WM_PSD_GREEKTEXTRECT = 0x0404;
-		public const int WM_PSD_ENVSTAMPRECT = 0x0405;
-		public const int WM_PSD_YAFULLPAGERECT = 0x0406;
-		public const int WM_CAP_UNICODE_START = 0x0464;
-		public const int WM_CHOOSEFONT_SETLOGFONT = 0x0465;
-		public const int WM_CAP_SET_CALLBACK_ERRORW = 0x0466;
-		public const int WM_CHOOSEFONT_SETFLAGS = 0x0466;
-		public const int WM_CAP_SET_CALLBACK_STATUSW = 0x0467;
-		public const int WM_CAP_DRIVER_GET_NAMEW = 0x0470;
-		public const int WM_CAP_DRIVER_GET_VERSIONW = 0x0471;
-		public const int WM_CAP_FILE_SET_CAPTURE_FILEW = 0x0478;
-		public const int WM_CAP_FILE_GET_CAPTURE_FILEW = 0x0479;
-		public const int WM_CAP_FILE_SAVEASW = 0x047b;
-		public const int WM_CAP_FILE_SAVEDIBW = 0x047d;
-		public const int WM_CAP_SET_MCI_DEVICEW = 0x04a6;
-		public const int WM_CAP_GET_MCI_DEVICEW = 0x04a7;
-		public const int WM_CAP_PAL_OPENW = 0x04b4;
-		public const int WM_CAP_PAL_SAVEW = 0x04b5;
-		public const int WM_CPL_LAUNCH = 0x07e8;
-		public const int WM_CPL_LAUNCHED = 0x07e9;
-		public const int WM_APP = 0x8000;
-		public const int WM_RASDIALEVENT = 0xcccd;
-		public const int EM_GETSEL = 0x00b0;
-		public const int EM_SETSEL = 0x00b1;
-		public const int EM_GETRECT = 0x00b2;
-		public const int EM_SETRECT = 0x00b3;
-		public const int EM_SETRECTNP = 0x00b4;
-		public const int EM_SCROLL = 0x00b5;
-		public const int EM_LINESCROLL = 0x00b6;
-		public const int EM_SCROLLCARET = 0x00b7;
-		public const int EM_GETMODIFY = 0x00b8;
-		public const int EM_SETMODIFY = 0x00b9;
-		public const int EM_GETLINECOUNT = 0x00ba;
-		public const int EM_LINEINDEX = 0x00bb;
-		public const int EM_SETHANDLE = 0x00bc;
-		public const int EM_GETHANDLE = 0x00bd;
-		public const int EM_GETTHUMB = 0x00be;
-		public const int EM_LINELENGTH = 0x00c1;
-		public const int EM_REPLACESEL = 0x00c2;
-		public const int EM_SETFONT = 0x00c3;
-		public const int EM_GETLINE = 0x00c4;
-		public const int EM_LIMITTEXT = 0x00c5;
-		public const int EM_SETLIMITTEXT = 0x00c5;
-		public const int EM_CANUNDO = 0x00c6;
-		public const int EM_UNDO = 0x00c7;
-		public const int EM_FMTLINES = 0x00c8;
-		public const int EM_LINEFROMCHAR = 0x00c9;
-		public const int EM_SETWORDBREAK = 0x00ca;
-		public const int EM_SETTABSTOPS = 0x00cb;
-		public const int EM_SETPASSWORDCHAR = 0x00cc;
-		public const int EM_EMPTYUNDOBUFFER = 0x00cd;
-		public const int EM_GETFIRSTVISIBLELINE = 0x00ce;
-		public const int EM_SETREADONLY = 0x00cf;
-		public const int EM_SETWORDBREAKPROC = 0x00d0;
-		public const int EM_GETWORDBREAKPROC = 0x00d1;
-		public const int EM_GETPASSWORDCHAR = 0x00d2;
-		public const int EM_SETMARGINS = 0x00d3;
-		public const int EM_GETMARGINS = 0x00d4;
-		public const int EM_GETLIMITTEXT = 0x00d5;
-		public const int EM_POSFROMCHAR = 0x00d6;
-		public const int EM_CHARFROMPOS = 0x00d7;
-		public const int EM_SETIMESTATUS = 0x00d8;
-		public const int EM_GETIMESTATUS = 0x00d9;
-		public const int EM_CANPASTE = 0x0432;
-		public const int EM_DISPLAYBAND = 0x0433;
-		public const int EM_EXGETSEL = 0x0434;
-		public const int EM_EXLIMITTEXT = 0x0435;
-		public const int EM_EXLINEFROMCHAR = 0x0436;
-		public const int EM_EXSETSEL = 0x0437;
-		public const int EM_FINDTEXT = 0x0438;
-		public const int EM_FORMATRANGE = 0x0439;
-		public const int EM_GETCHARFORMAT = 0x043a;
-		public const int EM_GETEVENTMASK = 0x043b;
-		public const int EM_GETOLEINTERFACE = 0x043c;
-		public const int EM_GETPARAFORMAT = 0x043d;
-		public const int EM_GETSELTEXT = 0x043e;
-		public const int EM_HIDESELECTION = 0x043f;
-		public const int EM_PASTESPECIAL = 0x0440;
-		public const int EM_REQUESTRESIZE = 0x0441;
-		public const int EM_SELECTIONTYPE = 0x0442;
-		public const int EM_SETBKGNDCOLOR = 0x0443;
-		public const int EM_SETCHARFORMAT = 0x0444;
-		public const int EM_SETEVENTMASK = 0x0445;
-		public const int EM_SETOLECALLBACK = 0x0446;
-		public const int EM_SETPARAFORMAT = 0x0447;
-		public const int EM_SETTARGETDEVICE = 0x0448;
-		public const int EM_STREAMIN = 0x0449;
-		public const int EM_STREAMOUT = 0x044a;
-		public const int EM_GETTEXTRANGE = 0x044b;
-		public const int EM_FINDWORDBREAK = 0x044c;
-		public const int EM_SETOPTIONS = 0x044d;
-		public const int EM_GETOPTIONS = 0x044e;
-		public const int EM_FINDTEXTEX = 0x044f;
-		public const int EM_GETWORDBREAKPROCEX = 0x0450;
-		public const int EM_SETWORDBREAKPROCEX = 0x0451;
-		public const int EM_SETUNDOLIMIT = 0x0452;
-		public const int EM_REDO = 0x0454;
-		public const int EM_CANREDO = 0x0455;
-		public const int EM_GETUNDONAME = 0x0456;
-		public const int EM_GETREDONAME = 0x0457;
-		public const int EM_STOPGROUPTYPING = 0x0458;
-		public const int EM_SETTEXTMODE = 0x0459;
-		public const int EM_GETTEXTMODE = 0x045a;
-		public const int EM_AUTOURLDETECT = 0x045b;
-		public const int EM_GETAUTOURLDETECT = 0x045c;
-		public const int EM_SETPALETTE = 0x045d;
-		public const int EM_GETTEXTEX = 0x045e;
-		public const int EM_GETTEXTLENGTHEX = 0x045f;
-		public const int EM_SHOWSCROLLBAR = 0x0460;
-		public const int EM_SETTEXTEX = 0x0461;
-		public const int EM_SETPUNCTUATION = 0x0464;
-		public const int EM_GETPUNCTUATION = 0x0465;
-		public const int EM_SETWORDWRAPMODE = 0x0466;
-		public const int EM_GETWORDWRAPMODE = 0x0467;
-		public const int EM_SETIMECOLOR = 0x0468;
-		public const int EM_GETIMECOLOR = 0x0469;
-		public const int EM_SETIMEOPTIONS = 0x046a;
-		public const int EM_GETIMEOPTIONS = 0x046b;
-		public const int EM_CONVPOSITION = 0x046c;
-		public const int EM_SETLANGOPTIONS = 0x0478;
-		public const int EM_GETLANGOPTIONS = 0x0479;
-		public const int EM_GETIMECOMPMODE = 0x047a;
-		public const int EM_FINDTEXTW = 0x047b;
-		public const int EM_FINDTEXTEXW = 0x047c;
-		public const int EM_RECONVERSION = 0x047d;
-		public const int EM_SETIMEMODEBIAS = 0x047e;
-		public const int EM_GETIMEMODEBIAS = 0x047f;
-		public const int EM_SETBIDIOPTIONS = 0x04c8;
-		public const int EM_GETBIDIOPTIONS = 0x04c9;
-		public const int EM_SETTYPOGRAPHYOPTIONS = 0x04ca;
-		public const int EM_GETTYPOGRAPHYOPTIONS = 0x04cb;
-		public const int EM_SETEDITSTYLE = 0x04cc;
-		public const int EM_GETEDITSTYLE = 0x04cd;
-		public const int EM_OUTLINE = 0x04dc;
-		public const int EM_GETSCROLLPOS = 0x04dd;
-		public const int EM_SETSCROLLPOS = 0x04de;
-		public const int EM_SETFONTSIZE = 0x04df;
-		public const int EM_GETZOOM = 0x04e0;
-		public const int EM_SETZOOM = 0x04e1;
-		public const int EM_GETVIEWKIND = 0x04e2;
-		public const int EM_SETVIEWKIND = 0x04e3;
-		public const int EM_GETPAGE = 0x04e4;
-		public const int EM_SETPAGE = 0x04e5;
-		public const int EM_GETHYPHENATEINFO = 0x04e6;
-		public const int EM_SETHYPHENATEINFO = 0x04e7;
-		public const int EM_GETPAGEROTATE = 0x04eb;
-		public const int EM_SETPAGEROTATE = 0x04ec;
-		public const int EM_GETCTFMODEBIAS = 0x04ed;
-		public const int EM_SETCTFMODEBIAS = 0x04ee;
-		public const int EM_GETCTFOPENSTATUS = 0x04f0;
-		public const int EM_SETCTFOPENSTATUS = 0x04f1;
-		public const int EM_GETIMECOMPTEXT = 0x04f2;
-		public const int EM_ISIME = 0x04f3;
-		public const int EM_GETIMEPROPERTY = 0x04f4;
-		public const int EM_GETQUERYRTFOBJ = 0x050d;
-		public const int EM_SETQUERYRTFOBJ = 0x050e;
-		public const int ACM_OPENA = 0x0464;
-		public const int ACM_PLAY = 0x0465;
-		public const int ACM_STOP = 0x0466;
-		public const int ACM_OPENW = 0x0467;
-		public const int BFFM_SETSTATUSTEXTA = 0x0464;
-		public const int BFFM_ENABLEOK = 0x0465;
-		public const int BFFM_SETSELECTIONA = 0x0466;
-		public const int BFFM_SETSELECTIONW = 0x0467;
-		public const int BFFM_SETSTATUSTEXTW = 0x0468;
-		public const int BM_GETCHECK = 0x00f0;
-		public const int BM_SETCHECK = 0x00f1;
-		public const int BM_GETSTATE = 0x00f2;
-		public const int BM_SETSTATE = 0x00f3;
-		public const int BM_SETSTYLE = 0x00f4;
-		public const int BM_CLICK = 0x00f5;
-		public const int BM_GETIMAGE = 0x00f6;
-		public const int BM_SETIMAGE = 0x00f7;
-		public const int BM_SETDONTCLICK = 0x00f8;
-		public const int CBEM_INSERTITEMA = 0x0401;
-		public const int CBEM_SETIMAGELIST = 0x0402;
-		public const int CBEM_GETIMAGELIST = 0x0403;
-		public const int CBEM_GETITEMA = 0x0404;
-		public const int CBEM_SETITEMA = 0x0405;
-		public const int CBEM_GETCOMBOCONTROL = 0x0406;
-		public const int CBEM_GETEDITCONTROL = 0x0407;
-		public const int CBEM_SETEXSTYLE = 0x0408;
-		public const int CBEM_GETEXSTYLE = 0x0409;
-		public const int CBEM_GETEXTENDEDSTYLE = 0x0409;
-		public const int CBEM_HASEDITCHANGED = 0x040a;
-		public const int CBEM_INSERTITEMW = 0x040b;
-		public const int CBEM_SETITEMW = 0x040c;
-		public const int CBEM_GETITEMW = 0x040d;
-		public const int CBEM_SETEXTENDEDSTYLE = 0x040e;
-		public const int CDM_FIRST = 0x0464;
-		public const int CDM_GETSPEC = 0x0464;
-		public const int CDM_GETFILEPATH = 0x0465;
-		public const int CDM_GETFOLDERPATH = 0x0466;
-		public const int CDM_GETFOLDERIDLIST = 0x0467;
-		public const int CDM_SETCONTROLTEXT = 0x0468;
-		public const int CDM_HIDECONTROL = 0x0469;
-		public const int CDM_SETDEFEXT = 0x046a;
-		public const int CDM_LAST = 0x04c8;
-		public const int DDM_SETFMT = 0x0400;
-		public const int DDM_DRAW = 0x0401;
-		public const int DDM_CLOSE = 0x0402;
-		public const int DDM_BEGIN = 0x0403;
-		public const int DDM_END = 0x0404;
-		public const int DL_BEGINDRAG = 0x0485;
-		public const int DL_DRAGGING = 0x0486;
-		public const int DL_DROPPED = 0x0487;
-		public const int DL_CANCELDRAG = 0x0488;
-		public const int DM_GETDEFID = 0x0400;
-		public const int DM_SETDEFID = 0x0401;
-		public const int DM_REPOSITION = 0x0402;
-		public const int FM_GETFOCUS = 0x0600;
-		public const int FM_GETDRIVEINFOA = 0x0601;
-		public const int FM_GETSELCOUNT = 0x0602;
-		public const int FM_GETSELCOUNTLFN = 0x0603;
-		public const int FM_GETFILESELA = 0x0604;
-		public const int FM_GETFILESELLFNA = 0x0605;
-		public const int FM_REFRESH_WINDOWS = 0x0606;
-		public const int FM_RELOAD_EXTENSIONS = 0x0607;
-		public const int FM_GETDRIVEINFOW = 0x0611;
-		public const int FM_GETFILESELW = 0x0614;
-		public const int FM_GETFILESELLFNW = 0x0615;
-		public const int HKM_SETHOTKEY = 0x0401;
-		public const int HKM_GETHOTKEY = 0x0402;
-		public const int HKM_SETRULES = 0x0403;
-		public const int IE_GETINK = 0x0496;
-		public const int IE_MSGFIRST = 0x0496;
-		public const int IE_SETINK = 0x0497;
-		public const int IE_GETPENTIP = 0x0498;
-		public const int IE_SETPENTIP = 0x0499;
-		public const int IE_GETERASERTIP = 0x049a;
-		public const int IE_SETERASERTIP = 0x049b;
-		public const int IE_GETBKGND = 0x049c;
-		public const int IE_SETBKGND = 0x049d;
-		public const int IE_GETGRIDORIGIN = 0x049e;
-		public const int IE_SETGRIDORIGIN = 0x049f;
-		public const int IE_GETGRIDPEN = 0x04a0;
-		public const int IE_SETGRIDPEN = 0x04a1;
-		public const int IE_GETGRIDSIZE = 0x04a2;
-		public const int IE_SETGRIDSIZE = 0x04a3;
-		public const int IE_GETMODE = 0x04a4;
-		public const int IE_SETMODE = 0x04a5;
-		public const int IE_GETINKRECT = 0x04a6;
-		public const int IE_GETAPPDATA = 0x04b8;
-		public const int IE_SETAPPDATA = 0x04b9;
-		public const int IE_GETDRAWOPTS = 0x04ba;
-		public const int IE_SETDRAWOPTS = 0x04bb;
-		public const int IE_GETFORMAT = 0x04bc;
-		public const int IE_SETFORMAT = 0x04bd;
-		public const int IE_GETINKINPUT = 0x04be;
-		public const int IE_SETINKINPUT = 0x04bf;
-		public const int IE_GETNOTIFY = 0x04c0;
-		public const int IE_SETNOTIFY = 0x04c1;
-		public const int IE_GETRECOG = 0x04c2;
-		public const int IE_SETRECOG = 0x04c3;
-		public const int IE_GETSECURITY = 0x04c4;
-		public const int IE_SETSECURITY = 0x04c5;
-		public const int IE_GETSEL = 0x04c6;
-		public const int IE_SETSEL = 0x04c7;
-		public const int IE_DOCOMMAND = 0x04c8;
-		public const int IE_GETCOMMAND = 0x04c9;
-		public const int IE_GETCOUNT = 0x04ca;
-		public const int IE_GETGESTURE = 0x04cb;
-		public const int IE_GETMENU = 0x04cc;
-		public const int IE_GETPAINTDC = 0x04cd;
-		public const int IE_GETPDEVENT = 0x04ce;
-		public const int IE_GETSELCOUNT = 0x04cf;
-		public const int IE_GETSELITEMS = 0x04d0;
-		public const int IE_GETSTYLE = 0x04d1;
-		public const int IPM_CLEARADDRESS = 0x0464;
-		public const int IPM_SETADDRESS = 0x0465;
-		public const int IPM_GETADDRESS = 0x0466;
-		public const int IPM_SETRANGE = 0x0467;
-		public const int IPM_SETFOCUS = 0x0468;
-		public const int IPM_ISBLANK = 0x0469;
-		public const int LVM_FIRST = 0x1000;
-		public const int LVM_GETBKCOLOR = 0x1000;
-		public const int LVM_SETBKCOLOR = 0x1001;
-		public const int LVM_GETIMAGELIST = 0x1002;
-		public const int LVM_SETIMAGELIST = 0x1003;
-		public const int LVM_GETITEMCOUNT = 0x1004;
-		public const int LVM_GETITEMA = 0x1005;
-		public const int LVM_SETITEMA = 0x1006;
-		public const int LVM_INSERTITEMA = 0x1007;
-		public const int LVM_DELETEITEM = 0x1008;
-		public const int LVM_DELETEALLITEMS = 0x1009;
-		public const int LVM_GETCALLBACKMASK = 0x100a;
-		public const int LVM_SETCALLBACKMASK = 0x100b;
-		public const int LVM_GETNEXTITEM = 0x100c;
-		public const int LVM_FINDITEMA = 0x100d;
-		public const int LVM_GETITEMRECT = 0x100e;
-		public const int LVM_SETITEMPOSITION = 0x100f;
-		public const int LVM_GETITEMPOSITION = 0x1010;
-		public const int LVM_GETSTRINGWIDTHA = 0x1011;
-		public const int LVM_HITTEST = 0x1012;
-		public const int LVM_ENSUREVISIBLE = 0x1013;
-		public const int LVM_SCROLL = 0x1014;
-		public const int LVM_REDRAWITEMS = 0x1015;
-		public const int LVM_ARRANGE = 0x1016;
-		public const int LVM_EDITLABELA = 0x1017;
-		public const int LVM_GETEDITCONTROL = 0x1018;
-		public const int LVM_GETCOLUMNA = 0x1019;
-		public const int LVM_SETCOLUMNA = 0x101a;
-		public const int LVM_INSERTCOLUMNA = 0x101b;
-		public const int LVM_DELETECOLUMN = 0x101c;
-		public const int LVM_GETCOLUMNWIDTH = 0x101d;
-		public const int LVM_SETCOLUMNWIDTH = 0x101e;
-		public const int LVM_GETHEADER = 0x101f;
-		public const int LVM_CREATEDRAGIMAGE = 0x1021;
-		public const int LVM_GETVIEWRECT = 0x1022;
-		public const int LVM_GETTEXTCOLOR = 0x1023;
-		public const int LVM_SETTEXTCOLOR = 0x1024;
-		public const int LVM_GETTEXTBKCOLOR = 0x1025;
-		public const int LVM_SETTEXTBKCOLOR = 0x1026;
-		public const int LVM_GETTOPINDEX = 0x1027;
-		public const int LVM_GETCOUNTPERPAGE = 0x1028;
-		public const int LVM_GETORIGIN = 0x1029;
-		public const int LVM_UPDATE = 0x102a;
-		public const int LVM_SETITEMSTATE = 0x102b;
-		public const int LVM_GETITEMSTATE = 0x102c;
-		public const int LVM_GETITEMTEXTA = 0x102d;
-		public const int LVM_SETITEMTEXTA = 0x102e;
-		public const int LVM_SETITEMCOUNT = 0x102f;
-		public const int LVM_SORTITEMS = 0x1030;
-		public const int LVM_SETITEMPOSITION32 = 0x1031;
-		public const int LVM_GETSELECTEDCOUNT = 0x1032;
-		public const int LVM_GETITEMSPACING = 0x1033;
-		public const int LVM_GETISEARCHSTRINGA = 0x1034;
-		public const int LVM_SETICONSPACING = 0x1035;
-		public const int LVM_SETEXTENDEDLISTVIEWSTYLE = 0x1036;
-		public const int LVM_GETEXTENDEDLISTVIEWSTYLE = 0x1037;
-		public const int LVM_GETSUBITEMRECT = 0x1038;
-		public const int LVM_SUBITEMHITTEST = 0x1039;
-		public const int LVM_SETCOLUMNORDERARRAY = 0x103a;
-		public const int LVM_GETCOLUMNORDERARRAY = 0x103b;
-		public const int LVM_SETHOTITEM = 0x103c;
-		public const int LVM_GETHOTITEM = 0x103d;
-		public const int LVM_SETHOTCURSOR = 0x103e;
-		public const int LVM_GETHOTCURSOR = 0x103f;
-		public const int LVM_APPROXIMATEVIEWRECT = 0x1040;
-		public const int LVM_SETWORKAREAS = 0x1041;
-		public const int LVM_GETSELECTIONMARK = 0x1042;
-		public const int LVM_SETSELECTIONMARK = 0x1043;
-		public const int LVM_SETBKIMAGEA = 0x1044;
-		public const int LVM_GETBKIMAGEA = 0x1045;
-		public const int LVM_GETWORKAREAS = 0x1046;
-		public const int LVM_SETHOVERTIME = 0x1047;
-		public const int LVM_GETHOVERTIME = 0x1048;
-		public const int LVM_GETNUMBEROFWORKAREAS = 0x1049;
-		public const int LVM_SETTOOLTIPS = 0x104a;
-		public const int LVM_GETITEMW = 0x104b;
-		public const int LVM_SETITEMW = 0x104c;
-		public const int LVM_INSERTITEMW = 0x104d;
-		public const int LVM_GETTOOLTIPS = 0x104e;
-		public const int LVM_FINDITEMW = 0x1053;
-		public const int LVM_GETSTRINGWIDTHW = 0x1057;
-		public const int LVM_GETCOLUMNW = 0x105f;
-		public const int LVM_SETCOLUMNW = 0x1060;
-		public const int LVM_INSERTCOLUMNW = 0x1061;
-		public const int LVM_GETITEMTEXTW = 0x1073;
-		public const int LVM_SETITEMTEXTW = 0x1074;
-		public const int LVM_GETISEARCHSTRINGW = 0x1075;
-		public const int LVM_EDITLABELW = 0x1076;
-		public const int LVM_GETBKIMAGEW = 0x108b;
-		public const int LVM_SETSELECTEDCOLUMN = 0x108c;
-		public const int LVM_SETTILEWIDTH = 0x108d;
-		public const int LVM_SETVIEW = 0x108e;
-		public const int LVM_GETVIEW = 0x108f;
-		public const int LVM_INSERTGROUP = 0x1091;
-		public const int LVM_SETGROUPINFO = 0x1093;
-		public const int LVM_GETGROUPINFO = 0x1095;
-		public const int LVM_REMOVEGROUP = 0x1096;
-		public const int LVM_MOVEGROUP = 0x1097;
-		public const int LVM_MOVEITEMTOGROUP = 0x109a;
-		public const int LVM_SETGROUPMETRICS = 0x109b;
-		public const int LVM_GETGROUPMETRICS = 0x109c;
-		public const int LVM_ENABLEGROUPVIEW = 0x109d;
-		public const int LVM_SORTGROUPS = 0x109e;
-		public const int LVM_INSERTGROUPSORTED = 0x109f;
-		public const int LVM_REMOVEALLGROUPS = 0x10a0;
-		public const int LVM_HASGROUP = 0x10a1;
-		public const int LVM_SETTILEVIEWINFO = 0x10a2;
-		public const int LVM_GETTILEVIEWINFO = 0x10a3;
-		public const int LVM_SETTILEINFO = 0x10a4;
-		public const int LVM_GETTILEINFO = 0x10a5;
-		public const int LVM_SETINSERTMARK = 0x10a6;
-		public const int LVM_GETINSERTMARK = 0x10a7;
-		public const int LVM_INSERTMARKHITTEST = 0x10a8;
-		public const int LVM_GETINSERTMARKRECT = 0x10a9;
-		public const int LVM_SETINSERTMARKCOLOR = 0x10aa;
-		public const int LVM_GETINSERTMARKCOLOR = 0x10ab;
-		public const int LVM_SETINFOTIP = 0x10ad;
-		public const int LVM_GETSELECTEDCOLUMN = 0x10ae;
-		public const int LVM_ISGROUPVIEWENABLED = 0x10af;
-		public const int LVM_GETOUTLINECOLOR = 0x10b0;
-		public const int LVM_SETOUTLINECOLOR = 0x10b1;
-		public const int LVM_CANCELEDITLABEL = 0x10b3;
-		public const int LVM_MAPINDEXTOID = 0x10b4;
-		public const int LVM_MAPIDTOINDEX = 0x10b5;
-		public const int LVM_ISITEMVISIBLE = 0x10b6;
-		public const int LVM_SETUNICODEFORMAT = 0x2005;
-		public const int LVM_GETUNICODEFORMAT = 0x2006;
-		public const int MCIWNDM_GETZOOM = 0x046d;
-		public const int MCIWNDM_REALIZE = 0x0476;
-		public const int MCIWNDM_SETTIMEFORMATA = 0x0477;
-		public const int MCIWNDM_GETTIMEFORMATA = 0x0478;
-		public const int MCIWNDM_VALIDATEMEDIA = 0x0479;
-		public const int MCIWNDM_PLAYTO = 0x047b;
-		public const int MCIWNDM_GETFILENAMEA = 0x047c;
-		public const int MCIWNDM_GETDEVICEA = 0x047d;
-		public const int MCIWNDM_GETPALETTE = 0x047e;
-		public const int MCIWNDM_SETPALETTE = 0x047f;
-		public const int MCIWNDM_GETERRORA = 0x0480;
-		public const int MCIWNDM_SETINACTIVETIMER = 0x0483;
-		public const int MCIWNDM_GETINACTIVETIMER = 0x0485;
-		public const int MCIWNDM_GET_SOURCE = 0x048c;
-		public const int MCIWNDM_PUT_SOURCE = 0x048d;
-		public const int MCIWNDM_GET_DEST = 0x048e;
-		public const int MCIWNDM_PUT_DEST = 0x048f;
-		public const int MCIWNDM_CAN_PLAY = 0x0490;
-		public const int MCIWNDM_CAN_WINDOW = 0x0491;
-		public const int MCIWNDM_CAN_RECORD = 0x0492;
-		public const int MCIWNDM_CAN_SAVE = 0x0493;
-		public const int MCIWNDM_CAN_EJECT = 0x0494;
-		public const int MCIWNDM_CAN_CONFIG = 0x0495;
-		public const int MCIWNDM_PALETTEKICK = 0x0496;
-		public const int MCIWNDM_NOTIFYMODE = 0x04c8;
-		public const int MCIWNDM_NOTIFYMEDIA = 0x04cb;
-		public const int MCIWNDM_NOTIFYERROR = 0x04cd;
-		public const int MCIWNDM_SETTIMEFORMATW = 0x04db;
-		public const int MCIWNDM_GETTIMEFORMATW = 0x04dc;
-		public const int MCIWNDM_GETFILENAMEW = 0x04e0;
-		public const int MCIWNDM_GETDEVICEW = 0x04e1;
-		public const int MCIWNDM_GETERRORW = 0x04e4;
-		public const int MSG_FTS_JUMP_VA = 0x0421;
-		public const int MSG_FTS_JUMP_QWORD = 0x0423;
-		public const int MSG_REINDEX_REQUEST = 0x0424;
-		public const int MSG_FTS_WHERE_IS_IT = 0x0425;
-		public const int MSG_GET_DEFFONT = 0x042d;
-		public const int NIN_SELECT = 0x0400;
-		public const int OCM__BASE = 0x2000;
-		public const int OCM_CTLCOLOR = 0x2019;
-		public const int OCM_DRAWITEM = 0x202b;
-		public const int OCM_MEASUREITEM = 0x202c;
-		public const int OCM_DELETEITEM = 0x202d;
-		public const int OCM_VKEYTOITEM = 0x202e;
-		public const int OCM_CHARTOITEM = 0x202f;
-		public const int OCM_COMPAREITEM = 0x2039;
-		public const int OCM_NOTIFY = 0x204e;
-		public const int OCM_COMMAND = 0x2111;
-		public const int OCM_HSCROLL = 0x2114;
-		public const int OCM_VSCROLL = 0x2115;
-		public const int OCM_CTLCOLORMSGBOX = 0x2132;
-		public const int OCM_CTLCOLOREDIT = 0x2133;
-		public const int OCM_CTLCOLORLISTBOX = 0x2134;
-		public const int OCM_CTLCOLORBTN = 0x2135;
-		public const int OCM_CTLCOLORDLG = 0x2136;
-		public const int OCM_CTLCOLORSCROLLBAR = 0x2137;
-		public const int OCM_CTLCOLORSTATIC = 0x2138;
-		public const int OCM_PARENTNOTIFY = 0x2210;
-		public const int PBM_SETRANGE = 0x0401;
-		public const int PBM_SETPOS = 0x0402;
-		public const int PBM_DELTAPOS = 0x0403;
-		public const int PBM_SETSTEP = 0x0404;
-		public const int PBM_STEPIT = 0x0405;
-		public const int PBM_SETRANGE32 = 0x0406;
-		public const int PBM_GETRANGE = 0x0407;
-		public const int PBM_GETPOS = 0x0408;
-		public const int PBM_SETBARCOLOR = 0x0409;
-		public const int PSM_SETCURSEL = 0x0465;
-		public const int PSM_REMOVEPAGE = 0x0466;
-		public const int PSM_ADDPAGE = 0x0467;
-		public const int PSM_CHANGED = 0x0468;
-		public const int PSM_RESTARTWINDOWS = 0x0469;
-		public const int PSM_REBOOTSYSTEM = 0x046a;
-		public const int PSM_CANCELTOCLOSE = 0x046b;
-		public const int PSM_QUERYSIBLINGS = 0x046c;
-		public const int PSM_UNCHANGED = 0x046d;
-		public const int PSM_APPLY = 0x046e;
-		public const int PSM_SETTITLEA = 0x046f;
-		public const int PSM_SETWIZBUTTONS = 0x0470;
-		public const int PSM_PRESSBUTTON = 0x0471;
-		public const int PSM_SETCURSELID = 0x0472;
-		public const int PSM_SETFINISHTEXTA = 0x0473;
-		public const int PSM_GETTABCONTROL = 0x0474;
-		public const int PSM_ISDIALOGMESSAGE = 0x0475;
-		public const int PSM_GETCURRENTPAGEHWND = 0x0476;
-		public const int PSM_INSERTPAGE = 0x0477;
-		public const int PSM_SETTITLEW = 0x0478;
-		public const int PSM_SETFINISHTEXTW = 0x0479;
-		public const int PSM_SETHEADERTITLEA = 0x047d;
-		public const int PSM_SETHEADERTITLEW = 0x047e;
-		public const int PSM_SETHEADERSUBTITLEA = 0x047f;
-		public const int PSM_SETHEADERSUBTITLEW = 0x0480;
-		public const int PSM_HWNDTOINDEX = 0x0481;
-		public const int PSM_INDEXTOHWND = 0x0482;
-		public const int PSM_PAGETOINDEX = 0x0483;
-		public const int PSM_INDEXTOPAGE = 0x0484;
-		public const int PSM_IDTOINDEX = 0x0485;
-		public const int PSM_INDEXTOID = 0x0486;
-		public const int PSM_GETRESULT = 0x0487;
-		public const int PSM_RECALCPAGESIZES = 0x0488;
-		public const int RB_INSERTBANDA = 0x0401;
-		public const int RB_DELETEBAND = 0x0402;
-		public const int RB_GETBARINFO = 0x0403;
-		public const int RB_SETBARINFO = 0x0404;
-		public const int RB_SETBANDINFOA = 0x0406;
-		public const int RB_SETPARENT = 0x0407;
-		public const int RB_HITTEST = 0x0408;
-		public const int RB_GETRECT = 0x0409;
-		public const int RB_INSERTBANDW = 0x040a;
-		public const int RB_SETBANDINFOW = 0x040b;
-		public const int RB_GETBANDCOUNT = 0x040c;
-		public const int RB_GETROWCOUNT = 0x040d;
-		public const int RB_GETROWHEIGHT = 0x040e;
-		public const int RB_IDTOINDEX = 0x0410;
-		public const int RB_GETTOOLTIPS = 0x0411;
-		public const int RB_SETTOOLTIPS = 0x0412;
-		public const int RB_SETBKCOLOR = 0x0413;
-		public const int RB_GETBKCOLOR = 0x0414;
-		public const int RB_SETTEXTCOLOR = 0x0415;
-		public const int RB_GETTEXTCOLOR = 0x0416;
-		public const int RB_SIZETORECT = 0x0417;
-		public const int RB_BEGINDRAG = 0x0418;
-		public const int RB_ENDDRAG = 0x0419;
-		public const int RB_DRAGMOVE = 0x041a;
-		public const int RB_GETBARHEIGHT = 0x041b;
-		public const int RB_GETBANDINFOW = 0x041c;
-		public const int RB_GETBANDINFOA = 0x041d;
-		public const int RB_MINIMIZEBAND = 0x041e;
-		public const int RB_MAXIMIZEBAND = 0x041f;
-		public const int RB_GETBANDBORDERS = 0x0422;
-		public const int RB_SHOWBAND = 0x0423;
-		public const int RB_SETPALETTE = 0x0425;
-		public const int RB_GETPALETTE = 0x0426;
-		public const int RB_MOVEBAND = 0x0427;
-		public const int RB_PUSHCHEVRON = 0x042b;
-		public const int SB_SETTEXTA = 0x0401;
-		public const int SB_GETTEXTA = 0x0402;
-		public const int SB_GETTEXTLENGTHA = 0x0403;
-		public const int SB_SETPARTS = 0x0404;
-		public const int SB_GETPARTS = 0x0406;
-		public const int SB_GETBORDERS = 0x0407;
-		public const int SB_SETMINHEIGHT = 0x0408;
-		public const int SB_SIMPLE = 0x0409;
-		public const int SB_GETRECT = 0x040a;
-		public const int SB_SETTEXTW = 0x040b;
-		public const int SB_GETTEXTLENGTHW = 0x040c;
-		public const int SB_GETTEXTW = 0x040d;
-		public const int SB_ISSIMPLE = 0x040e;
-		public const int SB_SETICON = 0x040f;
-		public const int SB_SETTIPTEXTA = 0x0410;
-		public const int SB_SETTIPTEXTW = 0x0411;
-		public const int SB_GETTIPTEXTA = 0x0412;
-		public const int SB_GETTIPTEXTW = 0x0413;
-		public const int SB_GETICON = 0x0414;
-		public const int SBM_SETPOS = 0x00e0;
-		public const int SBM_GETPOS = 0x00e1;
-		public const int SBM_SETRANGE = 0x00e2;
-		public const int SBM_GETRANGE = 0x00e3;
-		public const int SBM_ENABLE_ARROWS = 0x00e4;
-		public const int SBM_SETRANGEREDRAW = 0x00e6;
-		public const int SBM_SETSCROLLINFO = 0x00e9;
-		public const int SBM_GETSCROLLINFO = 0x00ea;
-		public const int SBM_GETSCROLLBARINFO = 0x00eb;
-		public const int SM_GETSELCOUNT = 0x07e8;
-		public const int SM_GETSERVERSELA = 0x07e9;
-		public const int SM_GETSERVERSELW = 0x07ea;
-		public const int SM_GETCURFOCUSA = 0x07eb;
-		public const int SM_GETCURFOCUSW = 0x07ec;
-		public const int SM_GETOPTIONS = 0x07ed;
-		public const int TAPI_REPLY = 0x0463;
-		public const int TB_ENABLEBUTTON = 0x0401;
-		public const int TB_CHECKBUTTON = 0x0402;
-		public const int TB_PRESSBUTTON = 0x0403;
-		public const int TB_HIDEBUTTON = 0x0404;
-		public const int TB_INDETERMINATE = 0x0405;
-		public const int TB_MARKBUTTON = 0x0406;
-		public const int TB_ISBUTTONENABLED = 0x0409;
-		public const int TB_ISBUTTONCHECKED = 0x040a;
-		public const int TB_ISBUTTONPRESSED = 0x040b;
-		public const int TB_ISBUTTONHIDDEN = 0x040c;
-		public const int TB_ISBUTTONINDETERMINATE = 0x040d;
-		public const int TB_ISBUTTONHIGHLIGHTED = 0x040e;
-		public const int TB_SETSTATE = 0x0411;
-		public const int TB_GETSTATE = 0x0412;
-		public const int TB_ADDBITMAP = 0x0413;
-		public const int TB_ADDBUTTONSA = 0x0414;
-		public const int TB_INSERTBUTTONA = 0x0415;
-		public const int TB_DELETEBUTTON = 0x0416;
-		public const int TB_GETBUTTON = 0x0417;
-		public const int TB_BUTTONCOUNT = 0x0418;
-		public const int TB_COMMANDTOINDEX = 0x0419;
-		public const int TB_SAVERESTOREA = 0x041a;
-		public const int TB_CUSTOMIZE = 0x041b;
-		public const int TB_ADDSTRINGA = 0x041c;
-		public const int TB_GETITEMRECT = 0x041d;
-		public const int TB_BUTTONSTRUCTSIZE = 0x041e;
-		public const int TB_SETBUTTONSIZE = 0x041f;
-		public const int TB_SETBITMAPSIZE = 0x0420;
-		public const int TB_AUTOSIZE = 0x0421;
-		public const int TB_GETTOOLTIPS = 0x0423;
-		public const int TB_SETTOOLTIPS = 0x0424;
-		public const int TB_SETPARENT = 0x0425;
-		public const int TB_SETROWS = 0x0427;
-		public const int TB_GETROWS = 0x0428;
-		public const int TB_GETBITMAPFLAGS = 0x0429;
-		public const int TB_SETCMDID = 0x042a;
-		public const int TB_CHANGEBITMAP = 0x042b;
-		public const int TB_GETBITMAP = 0x042c;
-		public const int TB_GETBUTTONTEXTA = 0x042d;
-		public const int TB_REPLACEBITMAP = 0x042e;
-		public const int TB_SETINDENT = 0x042f;
-		public const int TB_SETIMAGELIST = 0x0430;
-		public const int TB_GETIMAGELIST = 0x0431;
-		public const int TB_LOADIMAGES = 0x0432;
-		public const int TB_GETRECT = 0x0433;
-		public const int TB_SETHOTIMAGELIST = 0x0434;
-		public const int TB_GETHOTIMAGELIST = 0x0435;
-		public const int TB_SETDISABLEDIMAGELIST = 0x0436;
-		public const int TB_GETDISABLEDIMAGELIST = 0x0437;
-		public const int TB_SETSTYLE = 0x0438;
-		public const int TB_GETSTYLE = 0x0439;
-		public const int TB_GETBUTTONSIZE = 0x043a;
-		public const int TB_SETBUTTONWIDTH = 0x043b;
-		public const int TB_SETMAXTEXTROWS = 0x043c;
-		public const int TB_GETTEXTROWS = 0x043d;
-		public const int TB_GETOBJECT = 0x043e;
-		public const int TB_GETBUTTONINFOW = 0x043f;
-		public const int TB_SETBUTTONINFOW = 0x0440;
-		public const int TB_GETBUTTONINFOA = 0x0441;
-		public const int TB_SETBUTTONINFOA = 0x0442;
-		public const int TB_INSERTBUTTONW = 0x0443;
-		public const int TB_ADDBUTTONSW = 0x0444;
-		public const int TB_HITTEST = 0x0445;
-		public const int TB_SETDRAWTEXTFLAGS = 0x0446;
-		public const int TB_GETHOTITEM = 0x0447;
-		public const int TB_SETHOTITEM = 0x0448;
-		public const int TB_SETANCHORHIGHLIGHT = 0x0449;
-		public const int TB_GETANCHORHIGHLIGHT = 0x044a;
-		public const int TB_GETBUTTONTEXTW = 0x044b;
-		public const int TB_SAVERESTOREW = 0x044c;
-		public const int TB_ADDSTRINGW = 0x044d;
-		public const int TB_MAPACCELERATORA = 0x044e;
-		public const int TB_GETINSERTMARK = 0x044f;
-		public const int TB_SETINSERTMARK = 0x0450;
-		public const int TB_INSERTMARKHITTEST = 0x0451;
-		public const int TB_MOVEBUTTON = 0x0452;
-		public const int TB_GETMAXSIZE = 0x0453;
-		public const int TB_SETEXTENDEDSTYLE = 0x0454;
-		public const int TB_GETEXTENDEDSTYLE = 0x0455;
-		public const int TB_GETPADDING = 0x0456;
-		public const int TB_SETPADDING = 0x0457;
-		public const int TB_SETINSERTMARKCOLOR = 0x0458;
-		public const int TB_GETINSERTMARKCOLOR = 0x0459;
-		public const int TB_MAPACCELERATORW = 0x045a;
-		public const int TB_GETSTRINGW = 0x045b;
-		public const int TB_GETSTRINGA = 0x045c;
-		public const int TBM_GETPOS = 0x0400;
-		public const int TBM_GETRANGEMIN = 0x0401;
-		public const int TBM_GETRANGEMAX = 0x0402;
-		public const int TBM_GETTIC = 0x0403;
-		public const int TBM_SETTIC = 0x0404;
-		public const int TBM_SETPOS = 0x0405;
-		public const int TBM_SETRANGE = 0x0406;
-		public const int TBM_SETRANGEMIN = 0x0407;
-		public const int TBM_SETRANGEMAX = 0x0408;
-		public const int TBM_CLEARTICS = 0x0409;
-		public const int TBM_SETSEL = 0x040a;
-		public const int TBM_SETSELSTART = 0x040b;
-		public const int TBM_SETSELEND = 0x040c;
-		public const int TBM_GETPTICS = 0x040e;
-		public const int TBM_GETTICPOS = 0x040f;
-		public const int TBM_GETNUMTICS = 0x0410;
-		public const int TBM_GETSELSTART = 0x0411;
-		public const int TBM_GETSELEND = 0x0412;
-		public const int TBM_CLEARSEL = 0x0413;
-		public const int TBM_SETTICFREQ = 0x0414;
-		public const int TBM_SETPAGESIZE = 0x0415;
-		public const int TBM_GETPAGESIZE = 0x0416;
-		public const int TBM_SETLINESIZE = 0x0417;
-		public const int TBM_GETLINESIZE = 0x0418;
-		public const int TBM_GETTHUMBRECT = 0x0419;
-		public const int TBM_GETCHANNELRECT = 0x041a;
-		public const int TBM_SETTHUMBLENGTH = 0x041b;
-		public const int TBM_GETTHUMBLENGTH = 0x041c;
-		public const int TBM_SETTOOLTIPS = 0x041d;
-		public const int TBM_GETTOOLTIPS = 0x041e;
-		public const int TBM_SETTIPSIDE = 0x041f;
-		public const int TBM_SETBUDDY = 0x0420;
-		public const int TBM_GETBUDDY = 0x0421;
-		public const int TTM_ACTIVATE = 0x0401;
-		public const int TTM_SETDELAYTIME = 0x0403;
-		public const int TTM_ADDTOOLA = 0x0404;
-		public const int TTM_DELTOOLA = 0x0405;
-		public const int TTM_NEWTOOLRECTA = 0x0406;
-		public const int TTM_RELAYEVENT = 0x0407;
-		public const int TTM_GETTOOLINFOA = 0x0408;
-		public const int TTM_SETTOOLINFOA = 0x0409;
-		public const int TTM_HITTESTA = 0x040a;
-		public const int TTM_GETTEXTA = 0x040b;
-		public const int TTM_UPDATETIPTEXTA = 0x040c;
-		public const int TTM_GETTOOLCOUNT = 0x040d;
-		public const int TTM_ENUMTOOLSA = 0x040e;
-		public const int TTM_GETCURRENTTOOLA = 0x040f;
-		public const int TTM_WINDOWFROMPOINT = 0x0410;
-		public const int TTM_TRACKACTIVATE = 0x0411;
-		public const int TTM_TRACKPOSITION = 0x0412;
-		public const int TTM_SETTIPBKCOLOR = 0x0413;
-		public const int TTM_SETTIPTEXTCOLOR = 0x0414;
-		public const int TTM_GETDELAYTIME = 0x0415;
-		public const int TTM_GETTIPBKCOLOR = 0x0416;
-		public const int TTM_GETTIPTEXTCOLOR = 0x0417;
-		public const int TTM_SETMAXTIPWIDTH = 0x0418;
-		public const int TTM_GETMAXTIPWIDTH = 0x0419;
-		public const int TTM_SETMARGIN = 0x041a;
-		public const int TTM_GETMARGIN = 0x041b;
-		public const int TTM_POP = 0x041c;
-		public const int TTM_UPDATE = 0x041d;
-		public const int TTM_GETBUBBLESIZE = 0x041e;
-		public const int TTM_ADJUSTRECT = 0x041f;
-		public const int TTM_SETTITLEA = 0x0420;
-		public const int TTM_SETTITLEW = 0x0421;
-		public const int TTM_ADDTOOLW = 0x0432;
-		public const int TTM_DELTOOLW = 0x0433;
-		public const int TTM_NEWTOOLRECTW = 0x0434;
-		public const int TTM_GETTOOLINFOW = 0x0435;
-		public const int TTM_SETTOOLINFOW = 0x0436;
-		public const int TTM_HITTESTW = 0x0437;
-		public const int TTM_GETTEXTW = 0x0438;
-		public const int TTM_UPDATETIPTEXTW = 0x0439;
-		public const int TTM_ENUMTOOLSW = 0x043a;
-		public const int TTM_GETCURRENTTOOLW = 0x043b;
-		public const int UDM_SETRANGE = 0x0465;
-		public const int UDM_GETRANGE = 0x0466;
-		public const int UDM_SETPOS = 0x0467;
-		public const int UDM_GETPOS = 0x0468;
-		public const int UDM_SETBUDDY = 0x0469;
-		public const int UDM_GETBUDDY = 0x046a;
-		public const int UDM_SETACCEL = 0x046b;
-		public const int UDM_GETACCEL = 0x046c;
-		public const int UDM_SETBASE = 0x046d;
-		public const int UDM_GETBASE = 0x046e;
-		public const int UDM_SETRANGE32 = 0x046f;
-		public const int UDM_GETRANGE32 = 0x0470;
-		public const int UDM_SETPOS32 = 0x0471;
-		public const int UDM_GETPOS32 = 0x0472;
-		public const int UM_GETSELCOUNT = 0x07e8;
-		public const int UM_GETUSERSELA = 0x07e9;
-		public const int UM_GETUSERSELW = 0x07ea;
-		public const int UM_GETGROUPSELA = 0x07eb;
-		public const int UM_GETGROUPSELW = 0x07ec;
-		public const int UM_GETCURFOCUSA = 0x07ed;
-		public const int UM_GETCURFOCUSW = 0x07ee;
-		public const int UM_GETOPTIONS = 0x07ef;
-		public const int UM_GETOPTIONS2 = 0x07f0;
-		public const int WIZ_QUERYNUMPAGES = 0x040a;
-		public const int WIZ_NEXT = 0x040b;
-		public const int WIZ_PREV = 0x040c;
-		public const int WLX_WM_SAS = 0x0659;
-	}
-}
